Updated all dependencies + overall improvements
DSL element 'dexOptions' is obsolete and should be removed.
It will be removed in version 8.0 of the Android Gradle plugin.
Using it has no effect, and the AndroidGradle plugin optimizes dexing automatically.

Updated gradle to version 7
Replaced maven with maven-publish

Removed kotlin extensions as required and adapted mainActivity code to this

Removed multidex lib as not needed anymore from API 21

WebSocket lib old version was full of security bugs.

For Android projects consisting of an app with library dependencies, it is recommended to set checkDependencies to true
https://developer.android.com/studio/releases/gradle-plugin#improved_lint_for_library_dependencies